eslint and editorconfig
