+++ .eslintrc
@@ -1,25 +1,27 @@
 {
-  "extends": "eslint:recommended",
+  "extends": [
+    "eslint:recommended"
+  ],
+  "parserOptions": {
+    "ecmaVersion": 2017
+  },
+  "env": {
+    "commonjs": true,
+    "node": true,
+    "es6": true
+  },
   "globals": {
+    "window": true,
+    "imports": true,
     "ARGV": true,
-    "Int8Array": true,
-    "Int16Array": true,
-    "Int32Array": true,
-    "Intl": true,
-    "Map": true,
-    "Promise": true,
-    "Proxy": true,
-    "Set": true,
-    "Uint8Array": true,
-    "Uint8ClampedArray": true,
-    "Uint16Array": true,
-    "Uint32Array": true,
-    "WeakMap": true,
-    "WeakSet": true,
-    "imports": true
+    "print": true,
+    "printerr": true
   },
   "rules": {
     "no-case-declarations": 0,
-    "no-fallthrough": 0
+    "no-fallthrough": 0,
+    "no-console": 0,
+    "semi": [2, "always"],
+    "no-extra-semi": 0
   }
 }
