+++ package.json
@@ -1,10 +1,12 @@
 {
 	"private": true,
 	"scripts": {
-		"test": "lerna run test",
-		"postinstall": "lerna bootstrap"
+		"test": "lerna run test && npm run lint",
+		"postinstall": "lerna bootstrap",
+		"lint": "eslint ."
 	},
 	"devDependencies": {
+		"eslint": "^4.19.1",
 		"lerna": "^2.10.1"
 	}
 }
