+++ packages/os/test/index.js
@@ -1,3 +1,1 @@
-const os = require('../index');
-
-
+const os = require('../index'); // eslint-disable-line no-unused-vars
